Move followed artists import related methods to a specific service
